refactor: 폴더명 parser -> github 변경
